[8] Refactor
